2.Containers-VMs/1.2.Docker/2.5/app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9



